Refactor std::os to use sys::os
